fix(derive): move attributes builder trait
