+++ crates/derive/src/stages/test_utils/attributes_queue.rs
@@ -3,8 +3,8 @@
 use crate::{
     batch::SingleBatch,
     errors::{BuilderError, PipelineError, PipelineErrorKind, PipelineResult},
-    stages::attributes_queue::{AttributesBuilder, AttributesProvider},
-    traits::{OriginAdvancer, OriginProvider, ResettableStage},
+    stages::attributes_queue::AttributesProvider,
+    traits::{AttributesBuilder, OriginAdvancer, OriginProvider, ResettableStage},
 };
 use alloc::{boxed::Box, string::ToString, vec::Vec};
 use alloy_eips::BlockNumHash;
